Add more sophisticated "axum-like" routing
Lots of futzing with generics to allow handler functions of many
different signatures, whose arguments cause the API to automatically
extract and/or parse parts of the request, and whose return values cause
the API to automatically encode responses.

All of this still needs documentation and cleaning up/refactoring into
different modules.