Add bespoke xxdk.h (plus other misc. changes)
This is as opposed to the cgo-generated libxxdk.h.

Pros include:
- The ability to add const qualifiers to function arguments;
- Proper include guards;
- Cleaner public interface that hides a lot of irrelevant Go artifacts.

Cons include poorer maintainability; the bespoke header needs to be
manually updated to reflect changes to the Go code. However, I think
that the improvements in usability of the bindings are probably worth
it.

(It might be possible to generate the bespoke header from the generated
one, or at least to programatically check type signatures between them
for ABI compatibility; something to investigate.)

Other misc. changes in this commit:

- Add network health callbacks
- Move some common example functionality to common.cpp
- Start on connect_server example, to make sure the common xxdk
  functionality is all (or mostly) in place