Use TSS for error and info strings
Improves the error handling ergonomics of the library by removing the
requirement for the user to free errors.

This *should* be viable so long as Go functions called from C are
locked to the same OS thread for their duration. It seems that
in the current implementation of the Go runtime this is the case;
see comments and the implementation of `func cgocallbackg` here:
<https://cs.opensource.google/go/go/+/refs/tags/go1.24.0:src/runtime/cgocall.go>